+++ src/components/ui/index.ts
@@ -52,6 +52,9 @@ export { DatabaseError } from "./database-error";
 // Changelog Dialog
 export { ChangelogDialog } from "./changelog-dialog";
 
+// Update Notification
+export { UpdateNotification } from "./update-notification";
+
 // Pagination
 export { Pagination } from "./pagination";